+++ src/recommend.ts
@@ -1,6 +1,6 @@
 import * as vscode from 'vscode';
 
-import { deprecatedPackages, exists } from './analyzer';
+import { deprecatedPackages, exists, isGreaterOrEqual } from './analyzer';
 import { reviewCapacitorConfig } from './capacitor-configure';
 import { ionicBuild } from './ionic-build';
 import { ionicServe } from './ionic-serve';
@@ -30,6 +30,8 @@ import { getConfigurationName } from './build-configuration';
 import { liveReloadSSL } from './live-reload';
 import { npmInstall, npmUninstall } from './node-commands';
 import { writeIonic } from './extension';
+import { capacitorBuild } from './capacitor-build';
+import { getSetting, setSetting, WorkspaceSetting } from './workspace-state';
 
 export async function getRecommendations(
   project: Project,
@@ -56,7 +58,6 @@ export async function getRecommendations(
     project.add(runWeb);
     ionicState.runWeb = runWeb;
 
-    // project.add(new Tip('View In Editor', '', TipType.Run, 'Serve', undefined, 'Running on Web', `Project Served`).setAction(viewInEditor, 'http://localhost:8100'));
     const runPoints: Array<RunPoint> = [
       { text: 'Copying web assets', title: 'Copying...' },
       { text: 'ng run app:build', title: 'Building Web...' },
@@ -125,10 +126,14 @@ export async function getRecommendations(
         .setContextValue(Context.buildConfig)
         .canStop()
         .canAnimate()
-        .setTooltip('Builds the web project')
+        .setTooltip(
+          hasCapIos || hasCapAndroid
+            ? 'Builds the web project and copies to native platforms'
+            : 'Builds the web project'
+        )
     );
 
-    if (exists('@capacitor/core')) {
+    if (hasCapIos || hasCapAndroid) {
       project.add(
         new Tip('Sync', '', TipType.Sync, 'Capacitor Sync', undefined, 'Syncing', undefined)
           .setDynamicCommand(capacitorSync, project)
@@ -161,6 +166,25 @@ export async function getRecommendations(
           .setDynamicCommand(capacitorOpen, project, CapacitorPlatform.android)
           .setTooltip('Opens the native Android project in Android Studio')
       );
+    }
+    if (hasCapAndroid || hasCapIos) {
+      // cap build was added in v4.4.0
+      if (isGreaterOrEqual('@capacitor/core', '4.4.0')) {
+        project.add(
+          new Tip(
+            'Prepare Release Build',
+            '',
+            TipType.Build,
+            'Capacitor Build',
+            undefined,
+            'Preparing Release Build',
+            undefined
+          )
+            .setAction(capacitorBuild, project)
+            .canAnimate()
+            .setTooltip('Prepares native binaries suitable for uploading to the App Store or Play Store.')
+        );
+      }
     }
   }
 
@@ -296,7 +320,7 @@ function remoteLogging(project: Project): Tip {
 }
 
 function liveReload(): Tip {
-  const liveReload = vscode.workspace.getConfiguration('ionic').get('liveReload');
+  const liveReload = getSetting(WorkspaceSetting.liveReload);
   return new Tip('Live Reload', undefined, liveReload ? TipType.Check : TipType.Box, undefined)
     .setTooltip('Live reload will refresh the app whenever source code is changed.')
     .setAction(toggleLiveReload, liveReload)
@@ -305,7 +329,7 @@ function liveReload(): Tip {
 
 function useHttps(project: Project): Tip {
   if (!exists('@angular/core')) return;
-  const useHttps = vscode.workspace.getConfiguration('ionic').get('httpsForWeb');
+  const useHttps = getSetting(WorkspaceSetting.httpsForWeb);
   return new Tip('Use HTTPS', undefined, useHttps ? TipType.Check : TipType.Box, undefined)
     .setTooltip('Use HTTPS when running with web or Live Reload.')
     .setAction(toggleHttps, useHttps, project)
@@ -314,7 +338,7 @@ function useHttps(project: Project): Tip {
 
 function externalAddress(): Tip {
   if (!exists('@angular/core')) return;
-  const externalIP = vscode.workspace.getConfiguration('ionic').get('externalAddress');
+  const externalIP = getSetting(WorkspaceSetting.externalAddress);
   return new Tip('External Address', undefined, externalIP ? TipType.Check : TipType.Box, undefined)
     .setTooltip(
       'Using an external IP Address allows you to navigate to your application from other devices on the network.'
@@ -324,7 +348,7 @@ function externalAddress(): Tip {
 }
 
 function viewInEditor(): Tip {
-  const viewInEditor = vscode.workspace.getConfiguration('ionic').get('previewInEditor');
+  const viewInEditor = getSetting(WorkspaceSetting.previewInEditor);
   return new Tip('View In Editor', undefined, viewInEditor ? TipType.Check : TipType.Box, undefined)
     .setTooltip('Whether the app will be previewed in VS Code rather than a web browser')
     .setAction(toggleViewInEditor, viewInEditor)
@@ -339,11 +363,11 @@ function toggleRemoteLogging(project: Project, current: boolean): Promise<void> 
 }
 
 async function toggleLiveReload(current: boolean) {
-  await vscode.workspace.getConfiguration('ionic').update('liveReload', !current);
+  await setSetting(WorkspaceSetting.liveReload, !current);
 }
 
 async function toggleHttps(current: boolean, project: Project) {
-  await vscode.workspace.getConfiguration('ionic').update('httpsForWeb', !current);
+  await setSetting(WorkspaceSetting.httpsForWeb, !current);
   if (!current) {
     await showProgress('Enabling HTTPS', async () => {
       writeIonic('Installing @jcesarmobile/ssl-skip');
@@ -359,9 +383,9 @@ async function toggleHttps(current: boolean, project: Project) {
 }
 
 async function toggleExternalAddress(current: boolean) {
-  await vscode.workspace.getConfiguration('ionic').update('externalAddress', !current);
+  await setSetting(WorkspaceSetting.externalAddress, !current);
 }
 
 async function toggleViewInEditor(current: boolean) {
-  await vscode.workspace.getConfiguration('ionic').update('previewInEditor', !current);
+  await setSetting(WorkspaceSetting.previewInEditor, !current);
 }